+++ src/Plugins/MinIo/IMinIoClientFactory.cs
@@ -1,0 +1,17 @@
+﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
+// SPDX-License-Identifier: Apache License 2.0
+
+using Amazon.SecurityToken.Model;
+using Minio;
+
+namespace Monai.Deploy.Storage.MinIo
+{
+    public interface IMinIoClientFactory
+    {
+        MinioClient GetClient();
+
+        MinioClient GetClient(Credentials credentials);
+
+        MinioClient GetClient(Credentials credentials, string region);
+    }
+}
