+++ src/S3Policy/Policies/Condition.cs
@@ -1,0 +1,12 @@
+﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
+// SPDX-License-Identifier: Apache License 2.0
+
+namespace Monai.Deploy.Storage.S3Policy
+{
+    public class Condition
+    {
+        public StringLike? StringLike { get; set; }
+
+        public StringEquals? StringEquals { get; set; }
+    }
+}
