+++ src/Storage/API/IStorageService.cs
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache License 2.0
 
 using Amazon.SecurityToken.Model;
-using Monai.Deploy.Storage.Common;
 
-namespace Monai.Deploy.Storage
+namespace Monai.Deploy.Storage.API
 {
     public interface IStorageService
     {
@@ -21,7 +20,7 @@ namespace Monai.Deploy.Storage
         /// <param name="recursive">Whether to recurse into subdirectories</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns></returns>
-        IList<VirtualFileInfo> ListObjects(string bucketName, string prefix = "", bool recursive = false, CancellationToken cancellationToken = default);
+        Task<IList<VirtualFileInfo>> ListObjectsAsync(string bucketName, string prefix = "", bool recursive = false, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Downloads an objects as stream.
@@ -31,7 +30,7 @@ namespace Monai.Deploy.Storage
         /// <param name="callback">Action to be called when stream is ready</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task GetObject(string bucketName, string objectName, Action<Stream> callback, CancellationToken cancellationToken = default);
+        Task GetObjectAsync(string bucketName, string objectName, Action<Stream> callback, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Uploads an object.
@@ -44,7 +43,7 @@ namespace Monai.Deploy.Storage
         /// <param name="metadata">Metadata of the object</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task PutObject(string bucketName, string objectName, Stream data, long size, string contentType, Dictionary<string, string> metadata, CancellationToken cancellationToken = default);
+        Task PutObjectAsync(string bucketName, string objectName, Stream data, long size, string contentType, Dictionary<string, string> metadata, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Copies content of an object from source to destination.
@@ -55,7 +54,7 @@ namespace Monai.Deploy.Storage
         /// <param name="destinationObjectName">Name of the object in the destination bucket</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task CopyObject(string sourceBucketName, string sourceObjectName, string destinationBucketName, string destinationObjectName, CancellationToken cancellationToken = default);
+        Task CopyObjectAsync(string sourceBucketName, string sourceObjectName, string destinationBucketName, string destinationObjectName, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Removes an object.
@@ -64,7 +63,7 @@ namespace Monai.Deploy.Storage
         /// <param name="objectName">Name of the object in the bucket</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task RemoveObject(string bucketName, string objectName, CancellationToken cancellationToken = default);
+        Task RemoveObjectAsync(string bucketName, string objectName, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Removes a list of objects.
@@ -73,7 +72,7 @@ namespace Monai.Deploy.Storage
         /// <param name="objectNames">An enumerable of object names to be removed in the bucket</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task RemoveObjects(string bucketName, IEnumerable<string> objectNames, CancellationToken cancellationToken = default);
+        Task RemoveObjectsAsync(string bucketName, IEnumerable<string> objectNames, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Creates a folder with stub file.
@@ -82,7 +81,25 @@ namespace Monai.Deploy.Storage
         /// <param name="folderPath">Name/Path of the folder to be created. A stub file will also be created as this is required for MinIO</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task CreateFolder(string bucketName, string folderPath, CancellationToken cancellationToken = default);
+        Task CreateFolderAsync(string bucketName, string folderPath, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Verifies a list of artifacts to ensure that they exist.
+        /// </summary>
+        /// <param name="bucketName">Name of the bucket</param>
+        /// <param name="objectDict">Artifact dictionary to verify</param>
+        /// <returns>all valid artifacts</returns>
+        Task<Dictionary<string, string>> VerifyObjectsExistAsync(string bucketName, Dictionary<string, string> objectDict);
+
+        /// <summary>
+        /// Verifies the existence of an artifact to ensure that they exist.
+        /// </summary>
+        /// <param name="bucketName">Name of the bucket</param>
+        /// <param name="objectPair">Artifact to verify</param>
+        /// <returns>valid artifact</returns>
+        Task<KeyValuePair<string, string>> VerifyObjectExistsAsync(string bucketName, KeyValuePair<string, string> objectPair);
+
+        #region Temporary Credential APIs
 
         /// <summary>
         /// Creates temporary credentials for a specified folder for a specified length of time.
@@ -92,7 +109,7 @@ namespace Monai.Deploy.Storage
         /// <param name="durationSeconds">Expiration time of the credentials</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task<Credentials> CreateTemporaryCredentials(string bucketName, string folderName, int durationSeconds = 3600, CancellationToken cancellationToken = default);
+        Task<Credentials> CreateTemporaryCredentialsAsync(string bucketName, string folderName, int durationSeconds = 3600, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Copies content of an object from source to destination using temporary credentials.
@@ -104,7 +121,7 @@ namespace Monai.Deploy.Storage
         /// <param name="credentials">Temporary credentials used to connect</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task CopyObjectWithCredentials(string sourceBucketName, string sourceObjectName, string destinationBucketName, string destinationObjectName, Credentials credentials, CancellationToken cancellationToken = default);
+        Task CopyObjectWithCredentialsAsync(string sourceBucketName, string sourceObjectName, string destinationBucketName, string destinationObjectName, Credentials credentials, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Downloads an objects as stream using temporary credentials.
@@ -115,7 +132,7 @@ namespace Monai.Deploy.Storage
         /// <param name="callback">Action to be called when stream is ready</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task GetObjectWithCredentials(string bucketName, string objectName, Credentials credentials, Action<Stream> callback, CancellationToken cancellationToken = default);
+        Task GetObjectWithCredentialsAsync(string bucketName, string objectName, Credentials credentials, Action<Stream> callback, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Lists objects in a bucket using temporary credentials.
@@ -126,7 +143,7 @@ namespace Monai.Deploy.Storage
         /// <param name="recursive">Whether to recurse into subdirectories</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns></returns>
-        IList<VirtualFileInfo> ListObjectsWithCredentials(string bucketName, Credentials credentials, string? prefix = "", bool recursive = false, CancellationToken cancellationToken = default);
+        Task<IList<VirtualFileInfo>> ListObjectsWithCredentialsAsync(string bucketName, Credentials credentials, string prefix = "", bool recursive = false, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Uploads an object using temporary credentials.
@@ -140,7 +157,7 @@ namespace Monai.Deploy.Storage
         /// <param name="credentials">Temporary credentials used to connect</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task PutObjectWithCredentials(string bucketName, string objectName, Stream data, long size, string contentType, Dictionary<string, string> metadata, Credentials credentials, CancellationToken cancellationToken = default);
+        Task PutObjectWithCredentialsAsync(string bucketName, string objectName, Stream data, long size, string contentType, Dictionary<string, string> metadata, Credentials credentials, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Removes an object with temporary credentials.
@@ -150,7 +167,7 @@ namespace Monai.Deploy.Storage
         /// <param name="credentials">Temporary credentials used to connect</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task RemoveObjectWithCredentials(string bucketName, string objectName, Credentials credentials, CancellationToken cancellationToken = default);
+        Task RemoveObjectWithCredentialsAsync(string bucketName, string objectName, Credentials credentials, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Removes a list of objects with temporary credentials.
@@ -160,7 +177,7 @@ namespace Monai.Deploy.Storage
         /// <param name="credentials">Temporary credentials used to connect</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task RemoveObjectsWithCredentials(string bucketName, IEnumerable<string> objectNames, Credentials credentials, CancellationToken cancellationToken = default);
+        Task RemoveObjectsWithCredentialsAsync(string bucketName, IEnumerable<string> objectNames, Credentials credentials, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Creates a folder using temporary credentials.
@@ -170,22 +187,8 @@ namespace Monai.Deploy.Storage
         /// <param name="credentials">Temporary credentials used to connect</param>
         /// <param name="cancellationToken">Optional cancellation token. Defaults to default(CancellationToken)</param>
         /// <returns>Task</returns>
-        Task CreateFolderWithCredentials(string bucketName, string folderPath, Credentials credentials, CancellationToken cancellationToken = default);
+        Task CreateFolderWithCredentialsAsync(string bucketName, string folderPath, Credentials credentials, CancellationToken cancellationToken = default);
 
-        /// <summary>
-        /// Verifies a list of artifacts to ensure that they exist.
-        /// </summary>
-        /// <param name="bucketName">Name of the bucket</param>
-        /// <param name="objectDict">Artifact dictionary to verify</param>
-        /// <returns>all valid artifacts</returns>
-        Dictionary<string, string> VerifyObjectsExist(string bucketName, Dictionary<string, string> objectDict);
-
-        /// <summary>
-        /// Verifies the existance of an artifact to ensure that they exist.
-        /// </summary>
-        /// <param name="bucketName">Name of the bucket</param>
-        /// <param name="objectPair">Artifact to verify</param>
-        /// <returns>valid artifact</returns>
-        KeyValuePair<string, string> VerifyObjectExists(string bucketName, KeyValuePair<string, string> objectPair);
+        #endregion Temporary Credential APIs
     }
 }
