--- src/Storage/Common/Policies/Converter.cs
@@ -1,19 +1,0 @@
-﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
-// SPDX-License-Identifier: Apache License 2.0
-
-using Newtonsoft.Json;
-
-namespace Monai.Deploy.Storage.Common.Policies
-{
-    public partial class Policy
-    {
-        internal static class Converter
-        {
-            public static readonly JsonSerializerSettings Settings = new JsonSerializerSettings
-            {
-                MetadataPropertyHandling = MetadataPropertyHandling.Ignore,
-                DateParseHandling = DateParseHandling.None,
-            };
-        }
-    }
-}
