--- src/Storage/Common/Policies/StringEquals.cs
@@ -1,16 +1,0 @@
-﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
-// SPDX-License-Identifier: Apache License 2.0
-
-using Newtonsoft.Json;
-
-namespace Monai.Deploy.Storage.Common.Policies
-{
-    public class StringEquals
-    {
-        [JsonProperty(PropertyName = "s3:prefix")]
-        public string[]? S3Prefix { get; set; }
-
-        [JsonProperty(PropertyName = "s3:delimiter")]
-        public string[]? S3Delimiter { get; set; }
-    }
-}
