+++ src/Storage/Configuration/ConfigurationException.cs
@@ -12,11 +12,11 @@ namespace Monai.Deploy.Storage.Configuration
         {
         }
 
-        public ConfigurationException(string? message) : base(message)
+        public ConfigurationException(string message) : base(message)
         {
         }
 
-        public ConfigurationException(string? message, Exception? innerException) : base(message, innerException)
+        public ConfigurationException(string message, Exception innerException) : base(message, innerException)
         {
         }
